completato parsing redirezioni io e apici singoli e doppi. manca redirezione input doppia. test fatti, tutto funziona

+++ libft/ft_substr.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*ret;
-	size_t	i;
+	int	i;
 
 	if (s == 0)
 		return (0);
